final commit

+++ models/Job.cs
@@ -7,16 +7,16 @@ namespace WS381219_OOP
         // Tasks is a list of tasks associated with the job
         public List<Task> Tasks { get; set; }
         // Fitness is used to evaluate the quality of the job schedule
-        public int Fitness { get; set; }
+        public double Fitness { get; set; }
         // Stack to store jobs in correct task order
         private Stack<Task> jobStack;
 
-        // Constructor to initialize a job with a unique ID
+        // Constructor to initialise a job with a unique ID
         public Job(int jobID)
         {
             JobID = jobID;
             Tasks = new List<Task>();
-            // Initialize the jobStack here
+            // Initialise the jobStack here
             jobStack = new Stack<Task>(Tasks.Where(task => task.JobID == JobID));
         }
         // Method to add a task to the job
